Embed "ProjectReference" item rule for NuGet restore
C++/CLI projects use our NuGet restore/nomination code, without providing all our on-disk rules. The prior code required a non-embedded `ProjectReference` rule, which would work for .NET projects, but not for C++/CLI where the rule would not be found. Running with `set CPS_DiagnosticRuntime=1` would output to the "Project" pane of the Output window:

```
CPS Warning: 25816 :
Rule 'ProjectReference' was not resolved when it is subscribed through the ProjectJointSubscriptionService.
CPS Error: 25816 :
Generating snapshot for rule "ProjectReference" failed because the rule was missing from the project.
```

This commit creates a new `EvaluatedProjectReference` rule which _is_ embedded, and uses that in the NuGet restore code. The _Evaluated_ prefix is meant as an analog to _Collected_, where the latter represents items collected from a design-time build.